src/requirements.txt
black
flake8
ipython
isort
jupyter
jupyterlab
kedro
kedro-telemetry
mkgendocs
nbstripout
openpyxl
pytest-cov
pytest-mock
pytest
sphinx

#fix librairies
pathlib
ruamel-yaml

# Data Science
pandas
scikit-learn
kedro-mlflow

# Visualization
plotly
kedro-viz

# API
# fastapi[all]

